refactor: extract env schema module
